Change location of logger package
