+++ logger/logger.go
@@ -1,0 +1,11 @@
+package logger
+
+// Logger interface defines the behaviour of a log instance. Every logger
+// requires to implement these methods to be used by the Middleware
+type Logger interface {
+	Printf(format string, args ...interface{})
+	Debugf(format string, args ...interface{})
+	Infof(format string, args ...interface{})
+	Warnf(format string, args ...interface{})
+	Errorf(format string, args ...interface{})
+}
