+++ logger/middleware.go
@@ -1,4 +1,4 @@
-package tlog
+package logger
 
 import (
 	"io"
@@ -8,16 +8,6 @@ import (
 	"strings"
 	"sync"
 )
-
-// Logger interface defines the behaviour of a log instance. Every logger
-// requires to implement these methods to be used by the Middleware
-type Logger interface {
-	Printf(format string, args ...interface{})
-	Debugf(format string, args ...interface{})
-	Infof(format string, args ...interface{})
-	Warnf(format string, args ...interface{})
-	Errorf(format string, args ...interface{})
-}
 
 // TracePrefix is the prefix used to print a Terraform trace entry log
 const TracePrefix = "[TRACE] "
